Fix PHP 8.2 deprecations and WC 10.4 compat in combos module
- Declare 14 properties on WC_Combined_Item, 3 on WC_Product_Combo
  to eliminate ~1000+ dynamic property deprecation warnings per request
- Add WeakMap-based set/get/has_product_cart_prop() helpers on
  WC_LafkaCombos_Cart to avoid dynamic props on WC_Product objects
- Replace wc_enqueue_js() with wp_add_inline_script() in 5 files
  (deprecated since WC 10.4.0)
- Change float priority 9.9 to integer 9 (implicit conversion warning)
- Replace all (double) casts with (float) and (boolean) with (bool)
  across 8 files (40 occurrences) — deprecated in PHP 8.x
- Bump version to 8.2.6

+++ incl/combos/includes/admin/meta-boxes/class-lafka-combos-meta-box-product-data.php
@@ -1033,7 +1033,7 @@ class WC_LafkaCombos_Meta_Box_Product_Data {
 
 		$is_priced_individually  = isset( $item_data[ 'priced_individually' ] ) && 'yes' === $item_data[ 'priced_individually' ] ? 'yes' : '';
 		$is_shipped_individually = isset( $item_data[ 'shipped_individually' ] ) && 'yes' === $item_data[ 'shipped_individually' ] ? 'yes' : '';
-		$item_discount           = isset( $item_data[ 'discount' ] ) && (double) $item_data[ 'discount' ] > 0 ? $item_data[ 'discount' ] : '';
+		$item_discount           = isset( $item_data[ 'discount' ] ) && (float) $item_data[ 'discount' ] > 0 ? $item_data[ 'discount' ] : '';
 		$is_optional             = isset( $item_data[ 'optional' ] ) ? $item_data[ 'optional' ] : '';
 
 		// When adding a subscription-type product for the first time, enable "Priced Individually" by default.
